+++ P2/src/gameObjects/weapons/UCMMissile.java
@@ -8,19 +8,29 @@ import tp.p1.Game;
 public class UCMMissile extends Weapon {
 
 	public static final int DAMAGE = 1;
+	private Game game;
 
 	public UCMMissile(Game game, int row, int col) {
 		super(game, row, col, DAMAGE);
+		this.game = game;
+	}
 
+	public void checkDamage() {
+		GameObject go = game.isSomethingHere(row, col);
+		if (isAlive() && go != null && !go.equals(this) && !go.toString().equals(game.getUCMShip().toString())) {
+			performAttack(go);
+		}
 	}
 
 	@Override
 	public void move() {
-		if (row >= 0) {
+		checkDamage();
+		if (isAlive() && row >= 0) {
 			row--;
 			if (row < 0)
 				onDelete();
 		}
+		checkDamage();
 	}
 
 	@Override
@@ -51,9 +61,10 @@ public class UCMMissile extends Weapon {
 	}
 
 	@Override
-	protected GameObject parse(String stringFromFile, Game game, FileContentsVerifier verifier) throws FileContentsException {
+	protected GameObject parse(String stringFromFile, Game game, FileContentsVerifier verifier)
+			throws FileContentsException {
 		UCMMissile missile = null;
-		if(verifier.verifyWeaponString(stringFromFile, game) && stringFromFile.split(";")[0].equals("M")) {
+		if (verifier.verifyWeaponString(stringFromFile, game) && stringFromFile.split(";")[0].equals("M")) {
 			missile = new UCMMissile(game, row, col);
 			Weapon.game = game;
 			missile.row = getRowFromString(stringFromFile);
@@ -62,12 +73,10 @@ public class UCMMissile extends Weapon {
 		}
 		return missile;
 	}
-	
 
 	@Override
 	public int getNumber() {
 		return -1;
 	}
-
 
 }
